refactor: Extract version comparison from github_client.rs
Move UpdateInfo, PlatformInfo, and get_platform_target() into new
version_compare.rs module, keeping github_client.rs focused on HTTP
client logic and GitHub API interactions. Tests follow their respective
types into the new module.
